Tab system rewrite and restructure
