+++ src/components/Content/Editor/Editor.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import EditorTabs from "./EditorTabs.svelte";
-    import { hidden } from "./scripts/Tabs";
+    //import { hidden } from "./scripts/Tabs";
+    import { hidden } from "./scripts/Tab";
 </script>
 
 <EditorTabs />
