+++ src/components/FileTree/TreeViewNode.svelte
@@ -41,7 +41,8 @@
   export let icon = undefined;
 
   import { afterUpdate, getContext } from "svelte";
-  import { addTab } from "../Content/Editor/scripts/Tabs";
+  //import { addTab } from "../Content/Editor/scripts/Tabs";
+  import { addFileTab } from "../Content/Editor/scripts/Tab";
   import RenameModel from "../Modal/RenameModel.svelte";
   import LeafNodeMenu from "./LeafNodeMenu.svelte";
 
@@ -90,7 +91,7 @@
     clickNode(node);
   }}
   on:dblclick={async () => {
-    await addTab(path);
+    await addFileTab(path);
   }}
   on:keydown={(e) => {
     if (
